+++ Sources/Saga/ItemWriteMode.swift
@@ -1,3 +1,4 @@
+/// A value that represents item writing modes.
 public enum ItemWriteMode {
   /// This will write a file like `content/about.md` as `deploy/about.html`
   case keepAsFile
