+++ Sources/Saga/Path+Extensions.swift
@@ -2,27 +2,6 @@ import PathKit
 import Foundation
 
 public extension Path {
-  func makeOutputPath(itemWriteMode: ItemWriteMode) -> Path {
-    switch itemWriteMode {
-      case .keepAsFile:
-        return self.parent() + (self.lastComponentWithoutExtension.slugified + ".html")
-      case .moveToSubfolder:
-        return self.parent() + self.lastComponentWithoutExtension.slugified + "index.html"
-    }
-  }
-
-  func relativePath(from: Path) throws -> Path {
-    guard self.string.hasPrefix(from.string) else {
-      return self
-    }
-    let index = self.string.index(self.string.startIndex, offsetBy: from.string.count)
-    return Path(String(self.string[index...]).removingPrefix("/"))
-  }
-
-  var attributes: [FileAttributeKey : Any] {
-    return (try? FileManager.default.attributesOfItem(atPath: self.string)) ?? [:]
-  }
-
   var modificationDate: Date? {
     return self.attributes[.modificationDate] as? Date
   }
@@ -53,6 +32,27 @@ internal extension Path {
 
   func containsFile(named file: Path) -> Bool {
     return (self + file).isFile
+  }
+
+  func makeOutputPath(itemWriteMode: ItemWriteMode) -> Path {
+    switch itemWriteMode {
+      case .keepAsFile:
+        return self.parent() + (self.lastComponentWithoutExtension.slugified + ".html")
+      case .moveToSubfolder:
+        return self.parent() + self.lastComponentWithoutExtension.slugified + "index.html"
+    }
+  }
+
+  func relativePath(from: Path) throws -> Path {
+    guard self.string.hasPrefix(from.string) else {
+      return self
+    }
+    let index = self.string.index(self.string.startIndex, offsetBy: from.string.count)
+    return Path(String(self.string[index...]).removingPrefix("/"))
+  }
+
+  var attributes: [FileAttributeKey : Any] {
+    return (try? FileManager.default.attributesOfItem(atPath: self.string)) ?? [:]
   }
 }
