+++ Sources/Saga/Saga.swift
@@ -2,11 +2,35 @@ import Foundation
 import PathKit
 
 /// The main Saga class, used to configure and build your website.
+///
+/// ```swift
+/// @main
+/// struct Run {
+///   static func main() async throws {
+///     try await Saga(input: "content", output: "deploy", siteMetadata: EmptyMetadata())
+///       // All files in the input folder will be parsed to html, and written to the output folder.
+///       .register(
+///         metadata: EmptyMetadata.self,
+///         readers: [.parsleyMarkdownReader()],
+///         writers: [
+///           .itemWriter(swim(renderPage))
+///         ]
+///       )
+///
+///       // Run the step we registered above
+///       .run()
+///
+///       // All the remaining files that were not parsed to markdown, so for example images, raw html files and css,
+///       // are copied as-is to the output folder.
+///       .staticFiles()
+///   }
+/// }
+/// ```
 public class Saga<SiteMetadata: Metadata> {
   /// The root working path. This is automatically set to the same folder that holds `Package.swift`.
   public let rootPath: Path
 
-  /// The path that contains your text files, relative to the `rootPath`. For example "input".
+  /// The path that contains your text files, relative to the `rootPath`. For example "content".
   public let inputPath: Path
 
   /// The path that Saga will write the rendered website to, relative to the `rootPath`. For example "deploy".
@@ -41,6 +65,15 @@ public class Saga<SiteMetadata: Metadata> {
   }
 
   /// Register a new processing step.
+  ///
+  /// - Parameters:
+  ///   - folder: The folder (relative to `input`) to operate on. If `nil`, it operates on the `input` folder itself.
+  ///   - metadata: The metadata type used for the processing step. Required, but you can use ``EmptyMetadata`` if you don't need any custom metadata.
+  ///   - readers: The readers that will be used by this step.
+  ///   - itemWriteMode: The ``ItemWriteMode`` used by this step.
+  ///   - filter: A filter to only include certain items from the input folder.
+  ///   - writers: The writers that will be used by this step.
+  /// - Returns: The Saga instance itself, so you can chain further calls onto it.
   @discardableResult
   public func register<M: Metadata>(folder: Path? = nil, metadata: M.Type, readers: [Reader<M>], itemWriteMode: ItemWriteMode = .moveToSubfolder, filter: @escaping ((Item<M>) -> Bool) = { _ in true }, writers: [Writer<M, SiteMetadata>]) throws -> Self {
     let step = ProcessStep(folder: folder, readers: readers, filter: filter, writers: writers)
